Split out SlackDigestNotifier
